Imported using git-ubuntu import.

Publish parent: a7608ea4ab95152725d66353b7f6188d7d94b8d3
Changelog parent: a5f960aa8c4fb5f917e0fc0831552d3ce7d7a895
Unapplied parent: 4ae9e364d1325c8cde385468463a9c3958b76664

+++ sysdeps/mach/hurd/fchflags.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1991-2014 Free Software Foundation, Inc.
+/* Copyright (C) 1991-2016 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -25,7 +25,7 @@
 
 /* XXX should be __fchflags? */
 int
-fchflags (int fd, int flags)
+fchflags (int fd, unsigned long int flags)
 {
   error_t err;
